Revert "Merge pull request #60 from knowmetools/feature/require-verified-email"
This reverts commit 5d59142be8e8ff10c12608f53706b72ce98a53bd, reversing
changes made to d2385c5fbccb51f50217ff08ad0a65a945b1ab89.

As discussed in #70, we need to be able to tell the user if their
credentials are invalid or if their email is unverified. The cleanest
way to do this is to allow a user with an unverified email address to
authenticate, and then add an additional check for the unverified email.

Fixes #70