Add logic to handle sharing in filter backends.
Filters that were only used in a single view have been removed since
they just made the view logic harder to follow. The remaining filters
where combined into a single filter that covers all items with a
'km_user' as an owner that are sharable through the creation of a
'KMUserAccessor'.

Closes #173