Split profile serializers.
The serializers responsible for returning different parts of a profile
have been split into list and detail variants. The detail variants
include the same information as the list serializers with the next level
of the profile hierarchy included. For example, the profile topic detail
serializer includes the topic's items.

This change was made to avoid potential future performance issues
resulting from returning an entire profile in one call.